extract CLI arg parsing into shared cli_parse module
Pull split_top_level_commas, parse_cli_arg, parse_cli_arg_for_param,
and a new parse_cli_arg_as_list out of src/main.rs into a top-level
cli_parse module on the library crate.

The binary side keeps calling them via re-exports so existing call
sites stay byte-stable. The next commit needs to share the same
parsing logic with the AOT runtime helper that lives in src/vm/, and
having it on the library is the cleanest path.
